Make it possible to set a RAILS_ROOT explicitly.
This way, if your deployment is at /var/www/yourapp/current, where
current is a symlink, you can send HUP after a deployment, and it'll
load the new code, rather than the old code. Explicitly cd'ing to the
directory before starting unicorn means that the symlink is already
resolved and you have to USR2 to restart.